requirements.txt
jupyter
torch~=2.5.0
torchvision~=0.20.0
tqdm~=4.67.0
tensorboard
pandas~=2.2.3
wilds~=2.0.0
gdown~=5.2.0
ray~=2.39.0
matplotlib~=3.9.2
numpy~=2.1.2
PyYAML~=6.0.2
parcoords~=0.1.4
albumentations~=1.4.22
scikit-learn~=1.5.2
pillow~=11.0.0
scipy~=1.14.1
typing_extensions~=4.12.2
opencv-python~=4.10.0.84
medmnistc
wand > 0.6.10
